Fix unreliable memory initialization and kernel code structure
Using a shared memory location to wait for one core initializing
memory is unreliable, as cores don't hold any knowledge about the
initial memory state.
