+++ boards/qemu_rv32_virt_smp/src/threads/mod.rs
@@ -1,0 +1,3 @@
+pub mod main_thread;
+
+const NUM_THREADS: usize = 0;
